pyproject.toml
[tool.poetry]
name = "fastapi-realworld-example-app"
version = "0.0.0"
description = "Backend logic implementation for https://github.com/gothinkster/realworld with awesome FastAPI"
authors = ["Nik Sidnev <sidnev.nick@gmail.com>"]
license = "MIT"


[tool.poetry.dependencies]
python = "^3.7"

uvicorn = "^0.10.8"
fastapi = "^0.42.0"
pydantic = {version = "^0.32.2", extras = ["email"]}

passlib = {version = "^1.7", extras = ["bcrypt"]}

pyjwt = "^1.7"

asyncpg = "^0.18.3"
alembic = "^1.2"
databases = {version = "^0.2.5", extras = ["postgresql"]}

python-slugify = "^3.0"
Unidecode = "^1.1"

loguru = "^0.3.2"


[tool.poetry.dev-dependencies]
pre-commit = "^1.18"

black = {version = "^18.3-alpha.0", allows-prereleases = true}
isort = "^4.3"
autoflake = "^1.3"

wemake-python-styleguide = "^0.12.5"
mypy = "^0.720.0"
flake8-fixme = "^1.1"

pytest = "^5.1"
pytest-cov = "^2.7"
pytest-asyncio = "^0.10.0"
docker = "^4.0"
httpx = "^0.7.5"
asgi-lifespan = "^0.4.2"


[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
